requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
langchain-text-splitters
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv